+++ lizard_rainapp/layers.py
@@ -98,13 +98,8 @@ class RainAppAdapter(FewsJdbc):
     def layer(self, *args, **kwargs):
         """Return mapnik layers and styles."""
 
-        # We don't want to use shapes, unless configured.
-        # In the default case, the shapes cover all of the Netherlands
-        # with one tile per municipality, and that is too slow.
-        # However, for applications such as showing individual shapes
-        # within a single municipality, it can be turned on in settings.
-        if (not getattr(settings, 'RAINAPP_USE_SHAPES', False)
-            or not self.rainapp_config):
+        if not self.rainapp_config:
+            # Fall back to FEWSJDBC.
             return super(RainAppAdapter, self).layer(*args, **kwargs)
 
         slc = ShapeLegendClass.objects.get(descriptor=LEGEND_DESCRIPTOR)
@@ -169,10 +164,6 @@ class RainAppAdapter(FewsJdbc):
         return layers, styles
 
     def legend(self, updates=None):
-
-        if not getattr(settings, 'RAINAPP_USE_SHAPES', False):
-            return super(RainAppAdapter, self).legend(updates)
-
         slc = ShapeLegendClass.objects.get(descriptor=LEGEND_DESCRIPTOR)
         from lizard_shape.layers import AdapterShapefile
         la = {